Move list and create add todo form component
